Commit Cau 2
Viết hàm kiểm tra chuỗi Palindrome
